requirements.txt
Django~=3.2.12
gunicorn==20.0.4
djangorestframework==3.12.2
pytest-django==4.0.0
faker==4.1.2
requests==2.27.1
newsapi-python==0.2.6
beautifulsoup4===4.11.1
drf-yasg===1.21.4